Fixed and reorganized tests in hibernate-jcache
