+++ hibernate-jcache/src/test/java/org/hibernate/orm/test/jcache/BaseFunctionalTest.java
@@ -4,28 +4,29 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.jcache.test;
+package org.hibernate.orm.test.jcache;
 
 import org.hibernate.engine.spi.SessionFactoryImplementor;
 
-import org.hibernate.testing.junit4.BaseUnitTestCase;
-import org.junit.After;
-import org.junit.Before;
+import org.hibernate.testing.orm.junit.BaseUnitTest;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 
 /**
  * @author Steve Ebersole
  */
-public abstract class BaseFunctionalTest extends BaseUnitTestCase {
+@BaseUnitTest
+public abstract class BaseFunctionalTest {
 	private SessionFactoryImplementor sessionFactory;
 
-	@Before
+	@BeforeEach
 	public void createSessionFactory() {
 		assert sessionFactory == null || sessionFactory.isClosed();
 		TestHelper.preBuildAllCaches();
 		sessionFactory = TestHelper.buildStandardSessionFactory();
 	}
 
-	@After
+	@AfterEach
 	public void releaseSessionFactory() {
 		if ( sessionFactory != null ) {
 			sessionFactory.close();
