+++ hibernate-jcache/src/test/java/org/hibernate/orm/test/jcache/JCacheClasspathConfigUriTest.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.test.cache.jcache.config;
+package org.hibernate.orm.test.jcache;
 
 import java.util.Map;
 
@@ -12,21 +12,23 @@ import org.hibernate.cache.jcache.ConfigSettings;
 import org.hibernate.cfg.Environment;
 
 import org.hibernate.testing.junit4.BaseNonConfigCoreFunctionalTestCase;
+import org.hibernate.orm.test.jcache.domain.Product;
 import org.junit.Test;
 
 import static org.hibernate.testing.transaction.TransactionUtil.doInHibernate;
 
-public class JCacheConfigRelativePathTest
+public class JCacheClasspathConfigUriTest
 		extends BaseNonConfigCoreFunctionalTestCase {
 
 	@Override
 	protected void addSettings(Map settings) {
 		settings.put( Environment.CACHE_REGION_FACTORY, "jcache" );
-		settings.put( ConfigSettings.CONFIG_URI, "/hibernate-config/ehcache/jcache-ehcache-config.xml" );
+		settings.put( ConfigSettings.CONFIG_URI, "classpath://hibernate-config/ehcache/jcache-ehcache-config.xml" );
 	}
 
+	@Override
 	protected Class<?>[] getAnnotatedClasses() {
-		return new Class<?>[] {
+		return new Class<?>[]{
 				Product.class
 		};
 	}
