+++ hibernate-jcache/src/test/java/org/hibernate/orm/test/jcache/MissingCacheStrategyTest.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.jcache.test;
+package org.hibernate.orm.test.jcache;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -15,6 +15,7 @@ import org.hibernate.cache.CacheException;
 import org.hibernate.cache.jcache.ConfigSettings;
 import org.hibernate.cache.spi.SecondLevelCacheLogger;
 import org.hibernate.engine.spi.SessionFactoryImplementor;
+import org.hibernate.orm.test.jcache.TestHelper;
 import org.hibernate.service.spi.ServiceException;
 
 import org.hibernate.testing.junit4.BaseUnitTestCase;
