+++ hibernate-jcache/src/test/java/org/hibernate/orm/test/jcache/domain/Item.java
@@ -4,10 +4,10 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.jcache.test.domain;
+package org.hibernate.orm.test.jcache.domain;
 
-public class UuidItem {
-	private String id;
+public class Item {
+	private Long id;
 	private String name;
 	private String description;
 
@@ -19,11 +19,11 @@ public class UuidItem {
 		this.description = description;
 	}
 
-	public String getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(String id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
