+++ hibernate-jcache/src/test/java/org/hibernate/orm/test/jcache/domain/PhoneNumber.java
@@ -5,7 +5,7 @@
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
 
-package org.hibernate.jcache.test.domain;
+package org.hibernate.orm.test.jcache.domain;
 
 import java.io.Serializable;
