+++ hibernate-jcache/src/test/java/org/hibernate/orm/test/jcache/domain/UuidItem.java
@@ -4,10 +4,10 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.jcache.test.domain;
+package org.hibernate.orm.test.jcache.domain;
 
-public class Item {
-	private Long id;
+public class UuidItem {
+	private String id;
 	private String name;
 	private String description;
 
@@ -19,11 +19,11 @@ public class Item {
 		this.description = description;
 	}
 
-	public Long getId() {
+	public String getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(String id) {
 		this.id = id;
 	}
