+++ hibernate-jcache/src/test/java/org/hibernate/orm/test/jcache/domain/VersionedItem.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.jcache.test.domain;
+package org.hibernate.orm.test.jcache.domain;
 
 public class VersionedItem extends Item {
 	private Long version;
