+++ hibernate-jcache/src/test/resources/hibernate-config/domain/HolidayCalendar.hbm.xml
@@ -11,7 +11,7 @@
 
 <hibernate-mapping>
 
-    <class name="org.hibernate.jcache.test.domain.HolidayCalendar" table="CALENDAR" lazy="false">
+    <class name="org.hibernate.orm.test.jcache.domain.HolidayCalendar" table="CALENDAR" lazy="false">
         <id name="id" column="CALENDAR_ID">
             <generator class="native"/>
         </id>
