+++ hibernate-jcache/src/test/resources/hibernate-config/domain/Item.hbm.xml
@@ -9,7 +9,7 @@
         "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
         "http://hibernate.sourceforge.net/hibernate-mapping-3.0.dtd">
 
-<hibernate-mapping package="org.hibernate.jcache.test.domain">
+<hibernate-mapping package="org.hibernate.orm.test.jcache.domain">
 
 
     <class name="Item" table="Items">
